remove dvc.scm and dvc.fs.git module; rename dvc.fs.scm to dvc.fs.git
