+++ dvc/repo/experiments/executor/ssh.py
@@ -17,9 +17,10 @@ from dvc.repo.experiments.base import (
 from .base import BaseExecutor
 
 if TYPE_CHECKING:
+    from scmrepo.git import Git
+
     from dvc.machine import MachineManager
     from dvc.repo import Repo
-    from dvc.scm.git import Git
 
 logger = logging.getLogger(__name__)
