+++ setup.cfg
@@ -37,9 +37,6 @@ install_requires =
     ply>=3.9
     colorama>=0.3.9
     configobj>=5.0.6
-    gitpython>3
-    dulwich>=0.20.23
-    pygit2>=1.5.0
     nanotime>=0.5.2
     pyasn1>=0.4.1
     voluptuous>=0.11.7
@@ -76,6 +73,7 @@ install_requires =
     aiohttp-retry>=2.4.5
     diskcache>=5.2.1
     jaraco.windows>=5.7.0; python_version < '3.8' and sys_platform == 'win32'
+    scmrepo==0.0.1
 
 [options.extras_require]
 # gssapi should not be included in all_remotes, because it doesn't have wheels
