+++ tests/func/test_external_repo.py
@@ -1,10 +1,11 @@
 import os
 from unittest.mock import ANY, patch
 
+from scmrepo.git import Git
+
 from dvc.external_repo import CLONES, external_repo
 from dvc.objects.stage import stage
 from dvc.objects.transfer import transfer
-from dvc.scm.git import Git
 from dvc.utils import relpath
 from dvc.utils.fs import makedirs, remove
 from tests.unit.fs.test_repo import make_subrepo
